Run UniversalGP as a separate process
The basic idea is that first, the training and test data is written to a numpy
file. Then UniversalGP is started as a separate process that reads the
numpy file, does training and then writes the results in a different
numpy file. That file is then read by the benchmark class.

The new implementation depends on a UniversalGP dataset called
`sensitive_from_numpy` which creates a dataset from a numpy file.

This should prevent the "too many open files" error.